# contract/openapi.yaml
openapi: 3.1.0
info:
  title: TrueFigure SDK API
  version: "1.0.0"
  description: >
    Measurement contract for AI value verification. Three planes: write (events),
    config (deployments, namespaces, parameters, labels, webhooks), read (live facts,
    finalized figures). The SDK witnesses facts; measurement happens server-side by an
    independent engine. Structural guarantees: no content fields (BR-001), no value
    assertions (witness-only), fees never depend on findings (BR-002; billing has no
    read path here), refusals are first-class responses (BR-005), ingestion is
    idempotent by natural key (BR-009), every read against customer systems is
    auditable (BR-020). All responses are JSON objects with the uniform envelope
    { data, meta, errors }.
servers:
  - url: https://api.truefigure.io
security:
  - bearerAuth: []
...
paths:
  /v1/events:batch:
    post:
      tags: [write-plane]
      operationId: ingestEvents
      summary: Single ingestion door for all event families (max 500/batch)
      description: >
        Idempotent by natural key per family; re-sends are no-ops (TF-EVT-005, informational).
        Header X-TrueFigure-Mode: test returns synchronous parse-echo objects per event
        (resolved identity, matched work item, action class, accepted/reason) and stores nothing.
        Production mode validates + appends; identity resolution and joins run asynchronously
        (inspect via /v1/events/{event_key}/status).
      parameters:
        - in: header
          name: X-TrueFigure-Mode
          schema: { type: string, enum: [production, test], default: production }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "../schemas/events.schema.json" }
      responses:
        "200":
          description: Per-item results (accepted / rejected(reason) / duplicate) or echo objects in test mode.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchResponse" }
        "413": { $ref: "#/components/responses/Err" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/events/{event_key}/status:
    get:
      tags: [write-plane]
      operationId: eventStatus
      summary: Ingestion inspection (production debugging without support tickets)
      parameters:
        - { in: path, name: event_key, required: true, schema: { type: string }, description: URL-safe natural key hash returned in batch results }
      responses:
        "200":
          description: "accepted | rejected(reason) | resolved | joined | excluded(reason)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventStatus" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: "Workspace/deployment-scoped API keys; test keys cannot write production (TF-AUTH-003)." }
...
  responses:
...
    Err:
      description: Uniform envelope with errors[] populated from the closed registry.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    RateLimited:
      description: TF-RATE-001 with retry_after seconds.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
  schemas:
    Envelope:
      type: object
      required: [data, meta, errors]
      properties:
        data: { type: [object, "null"] }
        meta:
          type: object
          required: [schema_version, request_id, as_of]
          properties:
            schema_version: { type: string }
            request_id: { type: string }
            as_of: { type: string, format: date-time }
            deployment_id: { type: string }
            next_cursor: { type: [string, "null"] }
        errors:
          type: array
          items: { $ref: "#/components/schemas/ErrorObject" }
    ErrorObject:
      type: object
      required: [code, message, retryable, request_id]
      properties:
        code: { type: string, description: "NP-<PLANE>-<NNN> from the closed registry" }
        message: { type: string }
        field_path: { type: [string, "null"], description: "JSON pointer to the offending field" }
        expected: { type: [string, "null"] }
        received: { type: [string, "null"] }
        retryable: { type: boolean }
        retry_after: { type: [integer, "null"] }
        doc_url: { type: string }
        request_id: { type: string }
    BatchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          description: Per-item, index-aligned with the request batch. Never all-or-nothing.
          items:
            type: object
            required: [index, status, event_key]
            properties:
              index: { type: integer }
              status: { type: string, enum: [accepted, rejected, duplicate] }
              event_key: { type: string, description: URL-safe natural-key hash for /status inspection }
              error: { $ref: "#/components/schemas/ErrorObject" }
              echo:
                type: object
                description: Test mode only.
                properties:
                  resolved_person: { type: [string, "null"] }
                  resolution_basis: { type: string, enum: [deterministic, heuristic, unresolved] }
                  matched_work_item: { type: [string, "null"] }
                  action_class: { type: string }
                  accepted: { type: boolean }
                  reason: { type: [string, "null"] }
    EventStatus:
      type: object
      required: [event_key, status]
      properties:
        event_key: { type: string }
        status: { type: string, enum: [accepted, rejected, resolved, joined, excluded] }
        reason: { type: [string, "null"] }
        resolved_person: { type: [string, "null"] }
        joined_work_item: { type: [string, "null"] }
...